Changed css to camel case, added folders to components in order to  contain css individual files
