+++ client/components/CarouselButton/CarouselButton.jsx
@@ -1,0 +1,13 @@
+import React, {useState, useEffect} from 'react';
+
+
+const CarouselButton = ({buttonNum, onPageNumberButtonClick, arrOfFours}) => {  
+  return (
+  <button className="pageNumberBtn" onClick={(e) => {
+    onPageNumberButtonClick(e, arrOfFours[buttonNum - 1], buttonNum - 1)  
+  }}>{buttonNum}</button>
+  )
+
+}
+
+export default CarouselButton;
