+++ client/components/Index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import App from './App.jsx';
+import App from './App/App.jsx';
 import ReactDOM from 'react-dom';
 
 ReactDOM.render(<App />, document.getElementById('reviews'));
