+++ package.json
@@ -61,12 +61,14 @@
     "@babel/preset-react": "^7.10.4",
     "babel-jest": "^26.3.0",
     "babel-loader": "^8.1.0",
+    "css-loader": "^4.3.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.5",
     "jest": "^26.4.2",
     "jest-environment-enzyme": "^7.1.2",
     "jest-enzyme": "^7.1.2",
     "react-test-renderer": "^16.13.1",
+    "style-loader": "^1.3.0",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12"
   }
